+++ js-api/src/helpers.ts
@@ -35,6 +35,7 @@ export interface FormulaLine {
  * Represents a core annotation region parameters.
  */
 export interface AnnotationRegion {
+  id?: string;
   type?: string;
   header?: string;
   headerColor?: string;